Move Vec2d from hitrecord.py to geometry.py
